微调

+++ bezierIndicator/src/main/java/com/lcm/bezierbottomIndicator/viewpager/ShadowTransformer.java
@@ -81,8 +81,8 @@ public class ShadowTransformer implements ViewPager.OnPageChangeListener, ViewPa
         // and the views weren't created yet
         if (currentCard != null) {
             if (mScalingEnabled) {
-                currentCard.setScaleX((1 + mScaleValue * (1 - realOffset)));
-                currentCard.setScaleY((1 + mScaleValue * (1 - realOffset)));
+                currentCard.setScaleX( (1 + mScaleValue * (1 - realOffset)));
+                currentCard.setScaleY((float) (1 + mScaleValue * (1 - realOffset)));
             }
             currentCard.setCardElevation((baseElevation + baseElevation
                     * (CardAdapter.MAX_ELEVATION_FACTOR - 1) * (1 - realOffset)));
@@ -94,8 +94,8 @@ public class ShadowTransformer implements ViewPager.OnPageChangeListener, ViewPa
         // was already destroyed or a fragment might not have been created yet
         if (nextCard != null) {
             if (mScalingEnabled) {
-                nextCard.setScaleX((1 + mScaleValue * (realOffset)));
-                nextCard.setScaleY((1 + mScaleValue * (realOffset)));
+                nextCard.setScaleX((float) (1 + mScaleValue * (realOffset)));
+                nextCard.setScaleY((float) (1 + mScaleValue * (realOffset)));
             }
             nextCard.setCardElevation((baseElevation + baseElevation
                     * (CardAdapter.MAX_ELEVATION_FACTOR - 1) * (realOffset)));
